Retail Data Analysis Project/requirements.txt
pandas==2.2.2
numpy==1.26.4
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
faker==20.1.0
matplotlib==3.8.2
seaborn==0.13.0
jinja2==3.1.2
plotly==5.18.0
openpyxl==3.1.2
schedule==1.2.0
pytest==7.4.3